refactor : fix all the problems which was mentioned on previous branch
