+++ src/components/ChallengeManagement/create/UserCardListToAdd.tsx
@@ -1,24 +1,7 @@
 // src/components/ChallengeManagement/create/UserCardListToAdd.tsx
 import React from "react";
 import UserCardWithAddButton from "./UserCardWithAddButton";
-
-export type User = {
-  id: string;
-  username: string;
-  imagePath: string;
-  bio: string;
-  followersCount?: number;
-  followingCount?: number;
-  doneChallengesCount?: number;
-};
-
-export interface UserCardListProps {
-  users: User[];
-  searchTerm: string;
-  onAddUser: (user: User) => void;
-  onDelete?: (id: string, username: string) => void;
-  disabled?: boolean; // NEW: disables all add buttons
-}
+import type { UserCardListProps } from "@/types/challengeElementsTypes";
 
 const UserCardListToAdd: React.FC<UserCardListProps> = ({
   users,
@@ -46,14 +29,9 @@ const UserCardListToAdd: React.FC<UserCardListProps> = ({
         ) : (
           filtered.map((user) => (
             <UserCardWithAddButton
-              key={user.id}
               id={user.id}
               username={user.username}
               imagePath={user.imagePath}
-              bio={user.bio}
-              followersCount={user.followersCount ?? 0}
-              followingCount={user.followingCount ?? 0}
-              doneChallengesCount={user.doneChallengesCount ?? 0}
               onDelete={() => {}} // placeholder
               isOwner={true}
               className="w-full"
