+++ src/components/ChallengeManagement/create/UserCardWithAddButton.tsx
@@ -1,22 +1,9 @@
 // src/components/ChallengeManagement/create/UserCardWithAddButton.tsx
 import React from "react";
 import { useNavigate } from "react-router-dom";
+import type { ChallengeUserCardProps } from "@/types/challengeElementsTypes";
 
-export interface UserCardProps {
-  id: string;
-  username: string;
-  imagePath: string;
-  bio: string;
-  followersCount: number;
-  followingCount: number;
-  doneChallengesCount: number;
-  onDelete: (id: string, username: string) => void;
-  onAdd?: () => void;                 // NEW
-  isOwner: boolean;
-  className: string;
-}
-
-const UserCardWithAddButton: React.FC<UserCardProps> = ({
+const UserCardWithAddButton: React.FC<ChallengeUserCardProps> = ({
   id,
   username,
   imagePath,
@@ -50,7 +37,7 @@ const UserCardWithAddButton: React.FC<UserCardProps> = ({
 
   return (
     <div
-      className={`flex items-center justify-between mb-2 p-2 w-88 sm:w-100 md:w-110 h-24 border border-black rounded-[7px] cursor-pointer ${className}`}
+      className={`flex items-center justify-between mb-2 p-2 w-88 sm:w-100 md:w-110 h-24 border border-black rounded-primary-radius cursor-pointer ${className}`}
       onClick={handleCardClick}
     >
       <div className="flex items-center space-x-4">
@@ -64,7 +51,7 @@ const UserCardWithAddButton: React.FC<UserCardProps> = ({
 
       {isOwner && (
         <button
-          className="bg-white border-primary border-2 px-1 py-1 rounded-[10px] text-primary text-sm font-medium hover:bg-primary-hover transition-colors"
+          className="bg-white border-primary border-2 px-1 py-1 rounded-primary-radius text-primary text-sm font-medium hover:bg-primary-hover transition-colors"
           onClick={handleAddClick}
         >
           اضافه کردن
