+++ src/components/ChallengeManagement/info/BackButtonAndMenu.tsx
@@ -21,7 +21,7 @@ const BackButtonAndMenu: React.FC<BackButtonAndMenuProps> = ({ onMenuClick }) =>
       </button>
       <button
         onClick={onMenuClick}
-        className="text-primary w-11 h-11 border-2 border-primary rounded-[12.5px] px-2 py-2 flex items-center justify-center hover:bg-orange-50"
+        className="text-primary w-11 h-11 border-2 border-primary rounded-secondary-radius px-2 py-2 flex items-center justify-center hover:bg-primary-hover"
       >
         <Menu className="w-full h-full text-primary" />
       </button>
