+++ src/data/mockChallenges.tsx
@@ -1,0 +1,23 @@
+export const mockChallenges = [
+    {
+      id: "1",
+      title: "Mountain Climb Challenge",
+      description: "Climb a mountain in under 12 hours.",
+      imageUrl:
+        "https://www.muchbetteradventures.com/magazine/content/images/size/w2000/2024/04/mount-everest-at-sunset.jpg",
+    },
+    {
+      id: "2",
+      title: "Desert Trek Challenge",
+      description: "Complete a 30km desert trek in a single day.",
+      imageUrl:
+        "https://www.travelandleisure.com/thmb/Desert-Dunes_Fancyview_gettyimages-1216349476-f180b973c62f4b74b10b87bb736d9e9d.jpg",
+    },
+    {
+      id: "3",
+      title: "Ocean Swim Challenge",
+      description: "Swim across an open water stretch of 2km.",
+      imageUrl:
+        "https://www.adventure-life.com/sites/default/files/styles/hero_mobile/public/hero_images/hero-the-coastline-of-new-zealand.jpg?itok=V6n33P__",
+    },
+  ];
