+++ src/data/mockUsers.tsx
@@ -1,0 +1,50 @@
+import type { UserProfile } from "@/types/userTypes";
+
+
+export const mockUsers: UserProfile[] = [
+    {
+      id: "1",
+      username: "Alice",
+      imagePath: "https://randomuser.me/api/portraits/women/1.jpg",
+      bio: "Passionate about climbing and adventure.",
+      followersCount: 120,
+      followingCount: 80,
+      doneChallengesCount: 5,
+    },
+    {
+      id: "4",
+      username: "Damon",
+      imagePath: "https://randomuser.me/api/portraits/men/3.jpg",
+      bio: "Passionate about climbing and adventure.",
+      followersCount: 120,
+      followingCount: 80,
+      doneChallengesCount: 5,
+    },
+    {
+      id: "5",
+      username: "ching chang chong",
+      imagePath: "https://randomuser.me/api/portraits/women/2.jpg",
+      bio: "Passionate about climbing and adventure.",
+      followersCount: 120,
+      followingCount: 80,
+      doneChallengesCount: 5,
+    },
+    {
+      id: "2",
+      username: "Bob",
+      imagePath: "https://randomuser.me/api/portraits/men/1.jpg",
+      bio: "Love hiking and nature.",
+      followersCount: 150,
+      followingCount: 100,
+      doneChallengesCount: 7,
+    },
+    {
+      id: "3",
+      username: "Charlie",
+      imagePath: "https://randomuser.me/api/portraits/men/2.jpg",
+      bio: "Fitness enthusiast and challenge seeker.",
+      followersCount: 180,
+      followingCount: 90,
+      doneChallengesCount: 6,
+    },
+  ];
