+++ src/types/dropListTypes.ts
@@ -1,0 +1,8 @@
+export type CustomSelectProps = {
+    label?: string;
+    name: string;
+    options: { value: string; label: string }[];
+    icon?: React.ReactNode;
+    onIconClick?: () => void;
+    width?: string;
+  };
