+++ src/types/searchBarTypes.ts
@@ -1,0 +1,4 @@
+export interface SearchBarProps {
+    searchTerm: string;
+    onSearchTermChange: (value: string) => void;
+  }
