+++ src/types/userCardListType.ts
@@ -8,3 +8,17 @@ export interface UserCardListProps {
     onDelete: (id: string, username: string) => void;
     isOwner: boolean;
   }
+
+
+  export interface UserCardProps {
+    id: string;
+    username: string;
+    imagePath: string;
+    bio: string;
+    followersCount: number;
+    followingCount: number;
+    doneChallengesCount: number;
+    onDelete: (id: string, username: string) => void;
+    isOwner: boolean;
+    className: string;
+  }
